--- -- SQLite.sql
@@ -1,2 +1,0 @@
--- SQLite
-SELECT id from TERMINALES where lugar='San José'
