--- Registro1.txt
@@ -1,1 +1,0 @@
-{'Cedula': 5712, 'Contrasenia': '12345', 'Tipo': 'Admin'}
